ProductService was created, BUG wil be fixed about Duplicate value

+++ src/main/java/com/project/ecommerce/dto/product/ProductSaveRequest.java
@@ -19,7 +19,7 @@ public class ProductSaveRequest {
 	private String description;
 	@NotBlank
 	private BigDecimal price;
-	private int unitsStock;
+	private long unitsStock;
 	@NotBlank
 	private int categoryId;
 	@NotBlank
